requirements.txt
cassandra-driver==3.25.0
astrapy==0.0.2
simplejson==3.17.2
pyyaml~=6.0
pytest
pytest-mock==3.7.0
flask~=2.1.1
jsonpickle
dataclasses-json
PyJWT==1.7.1
Werkzeug~=2.1.1